add CommandHandlerInterface
