+++ husarion_ugv_manager/include/husarion_ugv_manager/plugins/action/command_handler_node.hpp
@@ -12,27 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef HUSARION_UGV_MANAGER_PLUGINS_ACTION_EXECUTE_COMMAND_NODE_HPP_
-#define HUSARION_UGV_MANAGER_PLUGINS_ACTION_EXECUTE_COMMAND_NODE_HPP_
+#ifndef HUSARION_UGV_MANAGER_PLUGINS_ACTION_COMMAND_HANDLER_NODE_HPP_
+#define HUSARION_UGV_MANAGER_PLUGINS_ACTION_COMMAND_HANDLER_NODE_HPP_
 
-#include <chrono>
-#include <memory>
 #include <string>
 
-#include "behaviortree_cpp/action_node.h"
 #include "behaviortree_cpp/basic_types.h"
-#include "rclcpp/logger.hpp"
+
+#include "husarion_ugv_manager/plugins/action/command_handler_interface.hpp"
 
 namespace husarion_ugv_manager
 {
 
-class CommandHandler : public BT::StatefulActionNode
+class CommandHandler : public CommandHandlerInterface
 {
 public:
-  explicit CommandHandler(const std::string & name, const BT::NodeConfig & conf)
-  : BT::StatefulActionNode(name, conf)
+  CommandHandler(const std::string & name, const BT::NodeConfig & conf)
+  : CommandHandlerInterface(name, conf)
   {
-    this->logger_ = std::make_shared<rclcpp::Logger>(rclcpp::get_logger(name));
   }
 
   ~CommandHandler() = default;
@@ -46,35 +43,10 @@ public:
   }
 
 protected:
-  BT::NodeStatus onStart() override;
-  BT::NodeStatus onRunning() override;
-  void onHalted() override;
-
-  /**
-   * @brief Orders execution of a command in a child process.
-   *
-   * @param command The command to be executed.
-   * @return true if the command execution was ordered successfully, false otherwise.
-   */
-  bool ExecuteCommandInChildProcess(const std::string & command);
-
-  /**
-   * @brief Logs the output of the command execution.
-   *
-   * @return true if it was possible to read and log the command output, false otherwise.
-   */
-  bool ReadAndLogCommandOutput();
-
-  void KillChildProcess();
-  bool TimeoutExceeded();
-
-  int pipefd_[2];
-  pid_t m_child_pid_;
-  std::shared_ptr<rclcpp::Logger> logger_;
-  std::chrono::milliseconds timeout_ms_;
-  std::chrono::time_point<std::chrono::steady_clock> command_time_;
+  std::string GetCommand() override;
+  float GetTimeout() override;
 };
 
 }  // namespace husarion_ugv_manager
 
-#endif  // HUSARION_UGV_MANAGER_PLUGINS_ACTION_EXECUTE_COMMAND_NODE_HPP_
+#endif  // HUSARION_UGV_MANAGER_PLUGINS_ACTION_COMMAND_HANDLER_NODE_HPP_
